feat: create a animation module to card swiper animation
